Re-wrote the negative expressions handling to archieve near hand-made (i.e. optimum) quality of iptables firewall. Now, instead of the linked-list that was created for negative expressions, we match all positive expressions before the negatives and all the negatives are together in one chain. This also fixed possible performance problems due to the large number of chains and rules that the packets had to traverse in order to get matched (or not matched).
The fact that now positive rules are matched before negatives has also the
benefit that not all traffic has to be matched against the negatives. Now,
first we select what might be good for a rule, and then we check if this
breaks the negative expressions.

Last, this made the iptables firewall much more clear and human readable.