Changed the method with which the current command line was saved for the error handler. The new way saved a fork() and gave some speed improvements.
Re-arranged the code so that the global variables and the services
definitions will be at the top of the file firehol.sh

Developed a work-around for a bug in iptables-save where:

! --uid-owner A

was saved as

--uid-owner !A

which was giving errors during restoration.
Unfortunatelly this made firehol saving the running firewall without the
use of /etc/init.d/iptables save.
The "save" argument to firehol saves the firewall to:

/etc/sysconfig/iptables