Various updates in 'helpme' mode. There is still one issue unsolved: When there are routes to specific nets befind a gateway, FireHOL checks if the IP of the host or the gateway are inside the networks routed through the gateway. This is wrong. In several cases, the network behind a gateway will not include the IP of the gateway or the IP of the interface the firehol host routes this traffic via.
However, there seems to be no easy solution to this. For PPP interfaces the
IP of the PPP concentrator is included in the IPs that are reported by
'ip route show dev ppp0' as the networks behind this device.

May be the 'scope' paramater of the routing table should be used to exclude
'scope link' routes. If 'scope link' routes are ommitted, the resulting set
could be used to identify the routes behind a router.