build: actually apply autoconf configuration
firehol cannot work properly if autoconf is not applied, not sure
why there is a complex mechanism to operate without it.

this logic tend to break as nobody actually test the process in
which actual users use this package.

I suggest to remove all these hacks, so developers will run make
after every change in scripts, or alternately generate a
configuration file that is sourced within the scripts instead of
substituting.

in the mean time, this fixes firehol script to apply the
AUTOCONF_RUN correctly within script, please notice that this
method of having .in.in file can also remove the commands.sed
usage as this exactly what autoconf is doing when processing
templates.

the same method should probably apply to other scripts.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>