Accept ICMP(v6) RELATED packets ASAP in interface
ICMP and ICMPv6 handling in the Linux kernel ensures all ICMP errors
will be marked as RELATED packets. These are important for the correct
functioning of other layers such as TCP (for instance Path MTU discovery).

Since 3.0.0, matches have been automatically added to firehol interfaces
and routers, so the user does not have to do anything special. It was
however possible to accidentally interfere with this by putting an
explicit drop of any packets towards the end of an interface.

This patch moves the ICMP RELATED match to the beginning of the firewall
(and similiarly the TCP-RESET RELATED match), so such a rule cannot
interfere with correct network operation.

At the same time, the ipv6error service has been marked as deprecated
since it has not done anything useful since 3.0.0.