Emit help in syslog on failure if no terminal
Addresses #163 where the user was left to their own devices after
a failure when loading via systemd and the information on stderr
was not captured anywhere they could find it.

We simply provide the necessary command to run by hand, so that whatever
output we expected them to see, they now will.