Fixed the time drift in fireqos when using sleepms.
Although the sleepms function in fireqos attempted to compensate for
variable CPU usage, the time-point at which the samples where taken,
was drifting. The reason for this was that the delay was calculated
relative to the previous sample time and not relative to the
application's start time.

This commit fixes this issue. The average drift is now 0ms, with a
maximum, absolute jitter of less than 6ms, on my system.

Verification of this could be done as follows:

1. Run the following for a few minutes (60 samples/min):
  $ fireqos status world-in |tai64n |tai64nlocal |tee fireqos.dat

2. Use the following awk command to get the min/max offsets:
  $ awk 'BEGIN{max=0; min=1000000000; sum=0; n=0} /TOTAL/ {for(i=0; \
  i<20; i++) {getline; n++; split($2,b,"."); sum+=b[2]; if (max<b[2]) \
  max=b[2]; if (min>b[2]) min=b[2];}; } END {avg=sum/n; \
  printf("num_samples=%d min_offset=%dus max_offset=%dus\n", \
  n, (avg-min)/1000, (max-avg)/1000)}' fireqos.dat