Moved the 'simple service' definitions into a central place.
Updating the 'simple service' definitions is not optimal, as it is
hard-coded in the firehol and fireqos scripts. This usually requires
an update in two different places.

This patch moves the 'simple service' definitions into the
sbin/simple_services.sh file. This requires firehol and fireqos to get
generated, as we probably do not want to add a run-time dependency on
them for including (actually sourcing) sbin/simple_services.sh.

In order to keep things easy for development, the firehol.sh and
fireqos.sh programs should be used. They include the external
definitions conditionally, so there's no need to regenerate (make)
firehol and fireqos constantly, while developing. The conditional
include will be skipped in the generated firehol and fireqos scripts.