Implemented simple_services.sh analog to functions.common and install.config.
This commit reverts the renaming and regenerating of the firehol and
fireqos scripts and includes simple_services.sh from within firehol
and fireqos.

Due to the fact that the definitions for document generation now
resides in firehol _and_ simple_services.sh, forced a change in the
doc/tools/mkservicelinks and doc/tools/mkserviceman scripts. These
scripts only accepted a single script file. The commit enhanced them to
handle more than one script, but required a change in the order of the
arguments that they get called with.